PRB/project_root/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
bcrypt
jose
authlib
psycopg2-binary
redis 
